+++ src/models/AbstractFactory/IShapeFactory.ts
@@ -2,5 +2,9 @@ import { Shape } from '../AbstractProduct/Shape';
 import { ShapeType } from '../enums/ShapeType';
 
 export interface IShapeFactory {
+    /**
+     * Hàm tạo shape
+     * @param type
+     */
     createShape(type: ShapeType): Shape;
 }
