+++ src/models/ConcreteProduct/Rectangle3D.ts
@@ -1,5 +1,6 @@
 import { Rectangle } from '../AbstractProduct/Rectangle';
 import { Point } from '../AbstractProduct/Point';
+import { Shape } from '../AbstractProduct/Shape';
 
 export class Rectangle3D extends Rectangle {
     constructor(location: Point = new Point(0, 0), width: number = 0, height: number = 0) {
@@ -8,5 +9,9 @@ export class Rectangle3D extends Rectangle {
 
     paint(ctx: CanvasRenderingContext2D): void {
         console.log('Rectangle3D');
+    }
+
+    clone(): Shape {
+        return new Rectangle3D();
     }
 }
